Renamed folder and exported module
